Move TemplateResolver into View namespace
