+++ src/Deprecator/Deprecator.php
@@ -10,8 +10,8 @@ use CraftCms\Cms\Element\Queries\Contracts\ElementQueryInterface;
 use CraftCms\Cms\Support\Arr;
 use CraftCms\Cms\Support\Str;
 use CraftCms\Cms\Support\Template;
-use CraftCms\Cms\Twig\TemplateResolver;
 use CraftCms\Cms\Twig\TwigExceptionMapper;
+use CraftCms\Cms\View\TemplateResolver;
 use Illuminate\Container\Attributes\Scoped;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Log;
