+++ src/Http/Controllers/Users/PhotoController.php
@@ -9,10 +9,10 @@ use CraftCms\Cms\Asset\Elements\Asset;
 use CraftCms\Cms\Element\Elements;
 use CraftCms\Cms\Http\RespondsWithFlash;
 use CraftCms\Cms\Support\File;
-use CraftCms\Cms\Twig\TemplateResolver;
 use CraftCms\Cms\User\Elements\User;
 use CraftCms\Cms\User\Users;
 use CraftCms\Cms\View\TemplateMode;
+use CraftCms\Cms\View\TemplateResolver;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
